app/requirements.txt
django==3.2.3
gunicorn==20.0.4
djangorestframework==3.12.4
python-decouple==3.4
astroid==2.11.5
autopep8==1.6.0
django-tastypie==0.14.4
djlint==1.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.11.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pylint==2.14.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyYAML==6.0
regex==2022.6.2
typing_extensions==4.2.0
